requirements.txt
pymongo
pandas
scikit-learn
matplotlib
joblib
python-dotenv